Move files plug_setting -> cat_settings.
